+++ heudiconv/tests/test_heuristics.py
@@ -64,10 +64,11 @@ def test_reproin_largely_smoke(tmpdir, heuristic, invocation):
             runner(args + ['--subjects', 'sub1', 'sub2'])
 
         if heuristic != 'reproin':
-            # none other heuristic has mighty infotoids atm
+            # if subject is not overriden, raise error
             with pytest.raises(NotImplementedError):
                 runner(args)
             return
+
     runner(args)
     ds = Dataset(str(tmpdir))
     assert ds.is_installed()